refactor: Split CLI into files
